requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jws==0.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
oauth2client==3.0.0
openapi-codec==1.3.2
Pillow==9.2.0
protobuf==4.21.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==1.5
python-decouple==3.6
python-jwt==2.0.1
pytz==2022.1
PyYAML==6.0
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
zipp==3.8.1
